requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
aom=3.5.0=h27087fc_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
blosc=1.21.5=hc2324a3_1
brotli=1.0.9=h166bdaf_9
brotli-bin=1.0.9=h166bdaf_9
brotli-python=1.0.9=py312h6a678d5_8
brunsli=0.1=h9c3ff4c_0
bzip2=1.0.8=h4bc722e_7
c-ares=1.33.1=heb4867d_0
ca-certificates=2024.7.4=hbcca054_0
cairo=1.16.0=hb05425b_5
certifi=2024.7.4=pyhd8ed1ab_0
cffi=1.17.0=py312h1671c18_0
cfitsio=3.470=hb418390_7
charls=2.2.0=h9c3ff4c_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.3.0=py312h68727a3_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.6.37=0
cuda-runtime=12.1.0=0
cuda-version=12.6=3
cycler=0.12.1=pyhd8ed1ab_0
dav1d=1.2.1=hd590300_0
dbus=1.13.18=hb2f20db_0
eigen=3.4.0=h00ab1b0_0
einops=0.8.0=pyhd8ed1ab_0
expat=2.6.2=h59595ed_0
ffmpeg=4.2.3=h167e202_0
filelock=3.15.4=pyhd8ed1ab_0
fontconfig=2.14.2=h14ed4e7_0
fonttools=4.53.1=py312h41a817b_0
freetype=2.12.1=h267a509_2
gdown=5.2.0=pypi_0
geos=3.12.2=he02047a_1
giflib=5.2.2=hd590300_0
glib=2.80.2=hf974151_0
glib-tools=2.80.2=hb6ce0ca_0
gmp=6.3.0=hac33072_2
gmpy2=2.1.5=py312h1d5cde6_1
gnutls=3.6.13=h85f3911_1
graphite2=1.3.14=h295c915_1
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h2=4.1.0=pyhd8ed1ab_0
harfbuzz=4.3.0=hf52aaf7_2
hdf5=1.12.1=nompi_h4df4325_104
hpack=4.0.0=pyh9f0ad1d_0
hyperframe=6.0.1=pyhd8ed1ab_0
icu=73.2=h59595ed_0
idna=3.8=pyhd8ed1ab_0
imagecodecs=2023.1.23=py312h81b8100_1
imageio=2.35.1=pyh12aca89_0
imgaug=0.4.0=pyhd8ed1ab_1
importlib-metadata=8.4.0=pyha770c72_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.4=pyhd8ed1ab_0
joblib=1.4.2=pyhd8ed1ab_0
jpeg=9e=h166bdaf_2
jxrlib=1.1=hd590300_3
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=py312h8572e83_1
krb5=1.20.1=h81ceb04_0
lame=3.100=h166bdaf_1003
lazy-loader=0.4=pyhd8ed1ab_1
lazy_loader=0.4=pyhd8ed1ab_1
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.40=hf3520f5_7
lerc=3.0=h9c3ff4c_0
libabseil=20240116.2=cxx17_he02047a_1
libaec=1.1.3=h59595ed_0
libavif=0.11.1=h8182462_2
libblas=3.9.0=1_h86c2bf4_netlib
libbrotlicommon=1.0.9=h166bdaf_9
libbrotlidec=1.0.9=h166bdaf_9
libbrotlienc=1.0.9=h166bdaf_9
libcblas=3.9.0=6_ha36c22a_netlib
libclang=14.0.6=default_h7634d5b_1
libclang13=14.0.6=default_h9986a30_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.11.0.15=0
libcups=2.3.3=h36d4200_3
libcurand=10.3.7.37=0
libcurl=8.1.2=h409715c_0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h0b41bf4_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=hd590300_2
libexpat=2.6.2=h59595ed_0
libffi=3.4.2=h7f98852_5
libgcc=14.1.0=h77fa898_1
libgcc-ng=14.1.0=h69a702a_1
libgfortran=14.1.0=h69a702a_1
libgfortran-ng=14.1.0=h69a702a_1
libgfortran5=14.1.0=hc5f4f2c_1
libglib=2.80.2=hf974151_0
libgomp=14.1.0=h77fa898_1
libhwloc=2.11.1=default_hecaa2ac_1000
libiconv=1.17=hd590300_2
libjpeg-turbo=2.0.0=h9bf148f_0
liblapack=3.9.0=6_ha36c22a_netlib
libllvm14=14.0.6=hcd5def8_4
libnghttp2=1.58.0=h47da74e_1
libnpp=12.0.2.50=0
libnsl=2.0.1=hd590300_0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.43=h2797004_0
libpq=12.17=hdbd6064_0
libprotobuf=4.25.3=h08a7969_0
libsqlite=3.46.0=hde9e2c9_0
libssh2=1.11.0=h0841786_0
libstdcxx=14.1.0=hc0a3c3a_1
libstdcxx-ng=14.1.0=h4852527_1
libtiff=4.5.1=h6a678d5_0
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.4.0=hd590300_0
libxcb=1.16=hb9d3cd8_1
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.7.0=h2c5496b_1
libxml2=2.12.7=hc051c1a_1
libzlib=1.2.13=h4ab18f5_6
libzopfli=1.0.3=h9c3ff4c_0
llvm-openmp=15.0.7=h0cdce71_0
lz4-c=1.9.4=hcb278e6_0
markupsafe=2.1.5=py312h98912ed_0
matplotlib-base=3.9.2=py312h854627b_0
mkl=2023.1.0=h213fc3f_46344
mpc=1.3.1=h24ddda3_0
mpfr=4.2.1=h38ae2d0_2
mpmath=1.3.0=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
mysql=5.7.20=hf484d3e_1001
ncurses=6.5=he02047a_1
nettle=3.6=he412f7d_0
networkx=3.3=pyhd8ed1ab_1
numpy=1.26.4=py312heda63a1_0
opencv=4.10.0=py312h907dce5_0
openh264=2.1.1=h780b84a_0
openjpeg=2.5.0=hfec8fc6_2
openssl=3.3.1=hb9d3cd8_3
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.2=py312h1d6d2e6_1
pcre2=10.43=hcad00b1_0
pillow=10.4.0=py312h5eee18b_0
pip=24.2=pyhd8ed1ab_0
pixman=0.43.2=h59595ed_0
positional-encodings=6.0.3=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
pycparser=2.22=pyhd8ed1ab_0
pyparsing=3.1.4=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.12.3=hab00c5b_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-tzdata=2024.1=pyhd8ed1ab_0
python_abi=3.12=5_cp312
pytorch=2.4.0=py3.12_cuda12.1_cudnn9.1.0_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pyhd8ed1ab_0
pywavelets=1.7.0=py312h4fc7734_0
pyyaml=6.0.2=py312h41a817b_0
qhull=2020.2=h434a139_5
qt-main=5.15.2=h53bd1ea_10
readline=8.2=h8228510_1
requests=2.32.3=pyhd8ed1ab_0
scikit-image=0.24.0=py312h1df14c2_2
scikit-learn=1.5.1=py312h775a589_0
scipy=1.14.1=py312h7d485d2_0
setuptools=72.2.0=pyhd8ed1ab_0
shapely=2.0.6=py312h8413631_0
six=1.16.0=pyh6c4a22f_0
snappy=1.2.1=ha2e4443_0
soupsieve=2.6=pypi_0
sqlite=3.46.0=h6d4b2fc_0
sympy=1.13.2=pypyh2585a3b_103
tbb=2021.12.0=h84d6215_4
threadpoolctl=3.5.0=pyhc1e730c_0
tifffile=2023.8.12=pyhd8ed1ab_0
tk=8.6.13=noxft_h4845f30_101
torchaudio=2.4.0=py312_cu121
torchtriton=3.0.0=py312
torchvision=0.19.0=py312_cu121
tqdm=4.66.5=pyhd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h8827d51_1
urllib3=2.2.2=pyhd8ed1ab_1
wheel=0.44.0=pyhd8ed1ab_0
x264=1!152.20180806=h14c3975_0
xkeyboard-config=2.42=h4ab18f5_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libx11=1.8.9=hb711507_1
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7f98852_2
zfp=1.0.1=hac33072_1
zipp=3.20.1=pyhd8ed1ab_0
zlib=1.2.13=h4ab18f5_6
zstandard=0.23.0=py312h3483029_0
zstd=1.5.6=ha6fb4c9_0
